+++ src/ui/about.rs
@@ -76,24 +76,44 @@ impl SimpleComponent for AboutDialog {
                 "简体中文 — Caibin Chen https://github.com/tigersoldier",
                 "日本語 — @zozonteq https://github.com/zozonteq",
                 // Hungarian?
+                "Svenska — Jakob Fridesjö https://github.com/jakobfridesjo",
             ].join("\n"),
 
             set_debug_info: &[
                 format!("Anime Launcher SDK: {SDK_VERSION}"),
                 format!("Anime Game Core: {CORE_VERSION}"),
                 String::new(),
-                format!("GTK: {}.{}.{}", gtk::major_version(), gtk::minor_version(), gtk::micro_version()),
+                format!("gtk: {}.{}.{}", gtk::major_version(), gtk::minor_version(), gtk::micro_version()),
                 format!("libadwaita: {}.{}.{}", adw::major_version(), adw::minor_version(), adw::micro_version()),
                 format!("pango: {}", gtk::pango::version_string()),
-                format!("cairo: {}", gtk::cairo::version_string()),
+                format!("cairo: {}", gtk::cairo::version_string())
             ].join("\n"),
 
             set_release_notes_version: &APP_VERSION,
             set_release_notes: &[
+                "<p>Added</p>",
+
+                "<ul>",
+                    "<li>Added new gamescope version compatibility</li>",
+                    "<li>Added \"launcher behavior\" option</li>",
+                    "<li>Added \"kill game process\" button when chosen behavior keeps launcher window open</li>",
+                    "<li>Bundled some icons into the app for consistency across the systems</li>",
+                    "<li>Added better panics handler</li>",
+                    "<li>Added Swedish</li>",
+                "</ul>",
+
                 "<p>Fixed</p>",
 
                 "<ul>",
-                    "<li>Added a workaround to remove outdated `Telemetry.dll` file which fixes login error 400 in game login screen</li>",
+                    "<li>Fixed predownload button sensitivity</li>",
+                "</ul>",
+
+                "<p>Changed</p>",
+
+                "<ul>",
+                    "<li>Improved pre-downloads state checking</li>",
+                    "<li>Replaced translation functions by `tr!` macro</li>",
+                    "<li>Reworked app resources structure</li>",
                 "</ul>",
             ].join("\n"),