+++ src/ui/preferences/general/mod.rs
@@ -11,15 +11,15 @@ use anime_launcher_sdk::anime_game_core::star_rail::consts::GameEdition;
 
 use anime_launcher_sdk::config::ConfigExt;
 use anime_launcher_sdk::star_rail::config::Config;
-use anime_launcher_sdk::star_rail::config::schema::launcher::LauncherStyle;
+use anime_launcher_sdk::star_rail::config::schema::launcher::{LauncherStyle, LauncherBehavior};
 
 pub mod components;
 
 use components::*;
 
-use super::main::PreferencesAppMsg;
-
 use crate::ui::migrate_installation::MigrateInstallationApp;
+use crate::ui::preferences::main::PreferencesAppMsg;
+
 use crate::i18n::*;
 use crate::*;
 
@@ -72,11 +72,11 @@ impl SimpleAsyncComponent for GeneralApp {
     view! {
         #[root]
         adw::PreferencesPage {
-            set_title: &tr("general"),
+            set_title: &tr!("general"),
             set_icon_name: Some("applications-system-symbolic"),
 
             add = &adw::PreferencesGroup {
-                set_title: &tr("appearance"),
+                set_title: &tr!("appearance"),
 
                 gtk::Box {
                     set_orientation: gtk::Orientation::Horizontal,
@@ -97,14 +97,14 @@ impl SimpleAsyncComponent for GeneralApp {
                             set_active: model.style == LauncherStyle::Modern,
 
                             gtk::Image {
-                                set_from_resource: Some("/org/app/images/modern.svg")
+                                set_resource: Some(&format!("{APP_RESOURCE_PATH}/images/modern.svg"))
                             },
 
                             connect_clicked => GeneralAppMsg::UpdateLauncherStyle(LauncherStyle::Modern)
                         },
 
                         gtk::Label {
-                            set_text: &tr("modern"),
+                            set_text: &tr!("modern"),
 
                             set_margin_top: 16
                         }
@@ -123,14 +123,14 @@ impl SimpleAsyncComponent for GeneralApp {
                             set_active: model.style == LauncherStyle::Classic,
 
                             gtk::Image {
-                                set_from_resource: Some("/org/app/images/classic.svg")
+                                set_resource: Some(&format!("{APP_RESOURCE_PATH}/images/classic.svg"))
                             },
 
                             connect_clicked => GeneralAppMsg::UpdateLauncherStyle(LauncherStyle::Classic)
                         },
 
                         gtk::Label {
-                            set_text: &tr("classic"),
+                            set_text: &tr!("classic"),
 
                             set_margin_top: 16
                         }
@@ -143,8 +143,8 @@ impl SimpleAsyncComponent for GeneralApp {
                 set_visible: model.style == LauncherStyle::Classic,
 
                 adw::ActionRow {
-                    set_title: &tr("update-background"),
-                    set_subtitle: &tr("update-background-description"),
+                    set_title: &tr!("update-background"),
+                    set_subtitle: &tr!("update-background-description"),
 
                     add_suffix = &gtk::Switch {
                         set_valign: gtk::Align::Center,
@@ -163,11 +163,11 @@ impl SimpleAsyncComponent for GeneralApp {
             },
 
             add = &adw::PreferencesGroup {
-                set_title: &tr("general"),
+                set_title: &tr!("general"),
 
                 adw::ComboRow {
-                    set_title: &tr("launcher-language"),
-                    set_subtitle: &tr("launcher-language-description"),
+                    set_title: &tr!("launcher-language"),
+                    set_subtitle: &tr!("launcher-language-description"),
 
                     set_model: Some(&gtk::StringList::new(&model.languages.iter().map(|lang| lang.as_str()).collect::<Vec<&str>>())),
 
@@ -193,11 +193,11 @@ impl SimpleAsyncComponent for GeneralApp {
                 },
 
                 adw::ComboRow {
-                    set_title: &tr("game-edition"),
+                    set_title: &tr!("game-edition"),
 
                     set_model: Some(&gtk::StringList::new(&[
-                        &tr("global"),
-                        &tr("china")
+                        &tr!("global"),
+                        &tr!("china")
                     ])),
 
                     set_selected: match CONFIG.launcher.edition {
@@ -230,8 +230,8 @@ impl SimpleAsyncComponent for GeneralApp {
                     set_margin_top: 16,
 
                     gtk::Button {
-                        set_label: &tr("migrate-installation"),
-                        set_tooltip_text: Some(&tr("migrate-installation-description")),
+                        set_label: &tr!("migrate-installation"),
+                        set_tooltip_text: Some(&tr!("migrate-installation-description")),
 
                         connect_clicked => GeneralAppMsg::OpenMigrateInstallation
                     }
@@ -239,10 +239,10 @@ impl SimpleAsyncComponent for GeneralApp {
             },
 
             add = &adw::PreferencesGroup {
-                set_title: &tr("status"),
+                set_title: &tr!("status"),
 
                 adw::ActionRow {
-                    set_title: &tr("game-version"),
+                    set_title: &tr!("game-version"),
 
                     add_suffix = &gtk::Label {
                         #[watch]
@@ -253,7 +253,7 @@ impl SimpleAsyncComponent for GeneralApp {
                                 VersionDiff::Diff { current, .. } |
                                 VersionDiff::Outdated { current, .. } => current.to_string(),
 
-                                VersionDiff::NotInstalled { .. } => tr("game-not-installed")
+                                VersionDiff::NotInstalled { .. } => tr!("game-not-installed")
                             }
 
                             None => String::from("?")
@@ -262,10 +262,10 @@ impl SimpleAsyncComponent for GeneralApp {
                         #[watch]
                         set_css_classes: match model.game_diff.as_ref() {
                             Some(diff) => match diff {
-                                VersionDiff::Latest { .. } => &["success"],
-                                VersionDiff::Predownload { .. } => &["accent"],
-                                VersionDiff::Diff { .. } => &["warning"],
-                                VersionDiff::Outdated { .. } => &["error"],
+                                VersionDiff::Latest { .. }       => &["success"],
+                                VersionDiff::Predownload { .. }  => &["accent"],
+                                VersionDiff::Diff { .. }         => &["warning"],
+                                VersionDiff::Outdated { .. }     => &["error"],
                                 VersionDiff::NotInstalled { .. } => &[]
                             }
 
@@ -276,17 +276,21 @@ impl SimpleAsyncComponent for GeneralApp {
                         set_tooltip_text: Some(&match model.game_diff.as_ref() {
                             Some(diff) => match diff {
                                 VersionDiff::Latest { .. } => String::new(),
-                                VersionDiff::Predownload { current, latest, .. } => tr_args("game-predownload-available", [
-                                    ("old", current.to_string().into()),
-                                    ("new", latest.to_string().into())
-                                ]),
-                                VersionDiff::Diff { current, latest, .. } => tr_args("game-update-available", [
-                                    ("old", current.to_string().into()),
-                                    ("new", latest.to_string().into())
-                                ]),
-                                VersionDiff::Outdated { latest, ..} => tr_args("game-outdated", [
-                                    ("latest", latest.to_string().into())
-                                ]),
+
+                                VersionDiff::Predownload { current, latest, .. } => tr!("game-predownload-available", {
+                                    "old" = current.to_string(),
+                                    "new" = latest.to_string()
+                                }),
+
+                                VersionDiff::Diff { current, latest, .. } => tr!("game-update-available", {
+                                    "old" = current.to_string(),
+                                    "new" = latest.to_string()
+                                }),
+
+                                VersionDiff::Outdated { latest, ..} => tr!("game-outdated", {
+                                    "latest" = latest.to_string()
+                                }),
+
                                 VersionDiff::NotInstalled { .. } => String::new()
                             }
 
@@ -296,8 +300,8 @@ impl SimpleAsyncComponent for GeneralApp {
                 },
 
                 adw::ActionRow {
-                    set_title: &tr("player-patch-version"),
-                    set_subtitle: &tr("player-patch-version-description"),
+                    set_title: &tr!("player-patch-version"),
+                    set_subtitle: &tr!("player-patch-version-description"),
 
                     add_suffix = &gtk::Label {
                         #[watch]
@@ -321,9 +325,9 @@ impl SimpleAsyncComponent for GeneralApp {
                         #[watch]
                         set_tooltip_text: Some(&match model.main_patch.as_ref() {
                             Some((_, status)) => match status {
-                                JadeitePatchStatusVariant::Unverified => tr("patch-testing-tooltip"),
-                                JadeitePatchStatusVariant::Broken => tr("patch-broken-tooltip"),
-                                JadeitePatchStatusVariant::Unsafe => tr("patch-unsafe-tooltip"),
+                                JadeitePatchStatusVariant::Unverified => tr!("patch-testing-tooltip"),
+                                JadeitePatchStatusVariant::Broken => tr!("patch-broken-tooltip"),
+                                JadeitePatchStatusVariant::Unsafe => tr!("patch-unsafe-tooltip"),
 
                                 _ => String::new()
                             }
@@ -335,11 +339,44 @@ impl SimpleAsyncComponent for GeneralApp {
             },
 
             add = &adw::PreferencesGroup {
-                set_title: &tr("options"),
+                set_title: &tr!("options"),
 
+                adw::ComboRow {
+                    set_title: &tr!("launcher-behavior"),
+                    set_subtitle: &tr!("launcher-behavior-description"),
+
+                    set_model: Some(&gtk::StringList::new(&[
+                        &tr!("nothing"),
+                        &tr!("hide", { "form" = "verb" }),
+                        &tr!("close", { "form" = "verb" }),
+                    ])),
+
+                    set_selected: match CONFIG.launcher.behavior {
+                        LauncherBehavior::Nothing => 0,
+                        LauncherBehavior::Hide    => 1,
+                        LauncherBehavior::Close   => 2
+                    },
+
+                    connect_selected_notify => |row| {
+                        if is_ready() {
+                            if let Ok(mut config) = Config::get() {
+                                config.launcher.behavior = [
+                                    LauncherBehavior::Nothing,
+                                    LauncherBehavior::Hide,
+                                    LauncherBehavior::Close
+                                ][row.selected() as usize];
+
+                                Config::update(config);
+                            }
+                        }
+                    }
+                }
+            },
+
+            add = &adw::PreferencesGroup {
                 adw::ActionRow {
-                    set_title: &tr("components"),
-                    set_subtitle: &tr("components-description"),
+                    set_title: &tr!("components"),
+                    set_subtitle: &tr!("components-description"),
 
                     add_suffix = &gtk::Image {
                         set_icon_name: Some("go-next-symbolic")
@@ -351,10 +388,10 @@ impl SimpleAsyncComponent for GeneralApp {
                 },
 
                 adw::ExpanderRow {
-                    set_title: &tr("wine-tools"),
+                    set_title: &tr!("wine-tools"),
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("command-line"),
+                        set_title: &tr!("command-line"),
                         set_subtitle: "wineconsole",
 
                         set_activatable: true,
@@ -363,7 +400,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     },
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("registry-editor"),
+                        set_title: &tr!("registry-editor"),
                         set_subtitle: "regedit",
 
                         set_activatable: true,
@@ -372,7 +409,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     },
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("explorer"),
+                        set_title: &tr!("explorer"),
                         set_subtitle: "explorer",
 
                         set_activatable: true,
@@ -381,7 +418,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     },
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("task-manager"),
+                        set_title: &tr!("task-manager"),
                         set_subtitle: "taskmgr",
 
                         set_activatable: true,
@@ -390,7 +427,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     },
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("configuration"),
+                        set_title: &tr!("configuration"),
                         set_subtitle: "winecfg",
 
                         set_activatable: true,
@@ -399,7 +436,7 @@ impl SimpleAsyncComponent for GeneralApp {
                     },
 
                     add_row = &adw::ActionRow {
-                        set_title: &tr("debugger"),
+                        set_title: &tr!("debugger"),
                         set_subtitle: "start winedbg",
 
                         set_activatable: true,
@@ -435,7 +472,7 @@ impl SimpleAsyncComponent for GeneralApp {
 
             style: CONFIG.launcher.style,
 
-            languages: SUPPORTED_LANGUAGES.iter().map(|lang| tr(format_lang(lang).as_str())).collect()
+            languages: SUPPORTED_LANGUAGES.iter().map(|lang| tr!(format_lang(lang).as_str())).collect()
         };
 
         let components_page = model.components_page.widget();
@@ -498,7 +535,7 @@ impl SimpleAsyncComponent for GeneralApp {
                         tracing::error!("Failed to download background picture");
 
                         sender.input(GeneralAppMsg::Toast {
-                            title: tr("background-downloading-failed"),
+                            title: tr!("background-downloading-failed"),
                             description: Some(err.to_string())
                         });
 
@@ -530,9 +567,9 @@ impl SimpleAsyncComponent for GeneralApp {
 
                     if let Err(err) = result {
                         sender.input(GeneralAppMsg::Toast {
-                            title: tr_args("wine-run-error", [
-                                ("executable", executable.join(" ").into())
-                            ]),
+                            title: tr!("wine-run-error", {
+                                "executable" = executable.join(" ")
+                            }),
                             description: Some(err.to_string())
                         });